+++ user_docs/en/changes.md
@@ -10,9 +10,14 @@ Any remaining users of SAPI4 speech synthesizers are encouraged to choose a more
 ### New Features
 
 * Add-on Store:
-  * Automatic update channels for add-ons can now be modified. (#3208)
-    * Automatic update channels can be selected for installed add-ons via an "Update channel" submenu.
-    * The default automatic update channel can be set from the Add-on Store panel in NVDA's settings.
+  * Automatic updates:
+    * Automatic update channels for add-ons can now be modified. (#3208)
+      * Automatic update channels can be selected for installed add-ons via an "Update channel" submenu.
+      * The default automatic update channel can be set from the Add-on Store panel in NVDA's settings.
+    * Automatic updates can now happen in the background.
+      * This can be enabled in the Add-on Store panel in NVDA's settings by changing "Automatic updates" to "Update Automatically". (#3208)
+    * Automatic updates can now update incompatible add-ons to another, newer, incompatible version.
+      * This can be enabled in the Add-on Store panel in NVDA's settings. (#3208)
   * Added an action to cancel the install of add-ons. (#15578, @hwf1324)
   * Added an action to retry the installation if the download/installation of an add-on fails. (#17090, @hwf1324)
   * The add-ons lists can be sorted by columns, including publication date, in ascending and descending order. (#15277, #16681, @nvdaes)